Initial checkin of libcon4m, which is expected to be the underlying support for the Con4m runtime.
Most of this code has been moved over from either nimutils or
libhatrack.

Note that we expect the con4m runtime to be fully pushed down to C,
which is part of the reason for this move. Additionally, I've mostly
migrated my `hatrack` library in here (currently, minus some of the
example programs and docs).

I've moved the build to the Meson build system, which is quite decent
to work with, especially compared to autoconf, CMake and build2. But
this move should make it much easier for all our dependent projects to
vendor the .a files they need too.

This initial build was only done on my M2 Mac. There are Linux
specific bits that I will work on next, particularly to ensure that
the hatrack bits compile cleanly with MUSL. However, I'm much more
confident in having a robust build across Mac and Linux now, whereas
the autoconf approach was like playing whack-a-mole. Once we went
MUSL, I could never get a build that worked in all cases...