Have a good chunk of the basic bits done for code generation, and am testing from the ground up (see tests/basic*.c4m for test programs so far). I've got a few things like assignment and indexing still to do, but those won't take long; the FFI will be the thing that is very important that takes a bit more time.
Added and changed instructions as needed; for some of the polymorphic
stuff, I'm not bothering w/ the polymorphism when it's a value type,
so I've added machine instructions for things like ADD.

I also added a temporary print instruction until FFI is done. Also I
am definitely going to be able to get rid of the need to carry the
type around for everything, so restructuring the stack / etc is a
matter of 'when' not 'if'. Also will simplify the VM thread's
interface a bit.