dev: make `dev run` exit non-zero when `c4test` exits non-zero
Before this commit, the new test workflow would indicate success when it
shouldn't (for example, when one of the tests produced a segfault).

This happened because `dev run` would exit with an exit code of 0 even
when `c4test` didn't. For now, make the `dev run` exit with the
exit code of `c4test`.

Note that we cannot currently add `set -e` in this script without making
other changes.