Reworking of lists, etc (#60)
n order to support more natural list ops, I made the list implementation use a read/write lock, and relegated the lock-free array to the cobwebs. I fixed some little bugs around it all.

I also refactored the collection algorithm to be breadth first instead of depth first, and re-arranged most structs to put pointers first, then 64-bit ints, since there were definitely some unaligned pointer issues w/ GCC (which could possibly persist in the stack layout, but hopefully not).

Debug builds now automatically run asan and ubsan; currently on my machine, there is one UB error, but as far as I can tell it is some false positive. What it's saying doesn't seem to map to the code even a little bit.

Particularly:

It's complaining about a bad boolean load.
The boolean is local, and explicitly initialized with bool flush = false.
The debugger agrees that the keyword argument that could modifies the address does not, and that, indeed, nothing modifies it on the call in question, when breaking on the function (verified w/ a hardware watchpoint).
This happens one and only one time across thousands and thousands of similar calls. But always in the same place.
HOWEVER, when the stack frame is entered, the value is somehow true already.
When I rearrange the declarations in the stack frame, I get the thing still.
I think there's a good chance there's just some weird ASAN byte being written, and it's just some weird false. But maybe not. Either way, I didn't chase it for long, and don't plan to (spent more time describing it than working it).

Also, along the way:
* Did all the con4m renaming I've been putting off, including the variadic macros, which I've mostly eliminated.
* String hash function has been customized to work right w/ u32, empty strings, etc. Also, fill in the list when there's an error in the lex phase.
* More basic tests
* Fix the comma in int formatting; the logic was sound but the counter never incremented.
* Made it so that uncaught exceptions will actually print (added a default handler for uncaught exceptions)
* Provide a value for TERM to the runner.
* Added a hard wrap to the terminal width for c4m_print(), which is currently on by default.