Fixed a bunch of little nits that fell out of the GC pointer work. There are still a few things that resort to a full scan that probably shouldn't.
Also, as part of trying to find all the little residual things, I
attempted to "fix" the macros for tree matching. There may not have
really been anything wrong with them, but moving from the typical
null-terminated varargs macros to the one that passes the # of varargs
on the stack has been problematic, and the code there isn't fully
working now.

So there are a few temporary functions to deal w/ the gaps.