HTTP support
Basic wrapping of libcurl to support PUT via c4m_http_upload(url:
string, ...) and GET via c4m_http_get(url: string, data: buffer, ...)

In good news, this supports TLS and gives us AWS v4 signature support
out of the box.

I did this on the path to getting modules fully working, but there are
several things that need to happen before it's fully on par w/ the
core capabilities in the old con4m:

1) Mime header support
2) Form posting
3) Certificate pinning.

This will be tested when I get to integrating it with modules.