Re-implemented the 'other' literal types:
Datetime
Time
Date
Size
Duration

URLs are done not as a separate type, but as a literal modifier that
calls CURL's URL validation (loose as it is).

Added an option to turn on / off static binding of calls exposed via
the FFI. If not set, it's on if glibc is off on linux, and off by
default on mac.

Also, for simple assignment ONLY, if the assignment is to a variable
with an explicitly declared type (and that type is a simple literal),
then we will infer the litmod type, picking the first one added
associated with the type.w

Some issue with the dynamic binding test in meson; just always leave static on for now.