Polished up the marshaling re-do, and then did a LOT of refactoring of data structures to make it easier to remove stuff not needed past compile time before we do a VM save.
There's definitely a bit more I'd like to do in that direction, mainly
not copying out functions and callbacks the way I currently do.

As part of that shuffling, I prepped everything to deal with
incremental loading of new modules, and did basic save/restore.  I've
got a bit more work to do there, but was at a good place for a
checkpoint.  basic04 is now an initial VM save/restore test as well.

I also aligned the GC implementation with the marshal implementation,
with a bit of prep for something I'll do shortly after getting back to
parity-- making the GC support multi-threading (I've decided its going
to be easiest to start with one central heap, and implement
cooperative migration).