Polished up the marshaling re-do, and then did a LOT of refactoring of data structures to make it easier to remove stuff not needed past compile time before we do a VM save.  There's definitely a bit more I'd like to do in that direction, mainly not copying out functions and callbacks the way I currently do.
As part of that shuffling, I prepped everything to deal with
incremental loading of new modules.

Then I went ahead and tested VM save/restore. It works, go figure.
Basic04 is now doing double duty as a basic test for it.

Need to test the incremental loading, but will do that tomorrow.